+++ server/handler.go
@@ -144,7 +144,7 @@ func (h *Handler) GetHotReadRegions() *statistics.StoreHotRegionInfos {
 }
 
 // GetHotBytesWriteStores gets all hot write stores stats.
-func (h *Handler) GetHotBytesWriteStores() map[uint64]uint64 {
+func (h *Handler) GetHotBytesWriteStores() map[uint64]float64 {
 	cluster := h.s.GetRaftCluster()
 	if cluster == nil {
 		return nil
@@ -153,7 +153,7 @@ func (h *Handler) GetHotBytesWriteStores() map[uint64]uint64 {
 }
 
 // GetHotBytesReadStores gets all hot write stores stats.
-func (h *Handler) GetHotBytesReadStores() map[uint64]uint64 {
+func (h *Handler) GetHotBytesReadStores() map[uint64]float64 {
 	cluster := h.s.GetRaftCluster()
 	if cluster == nil {
 		return nil
@@ -162,7 +162,7 @@ func (h *Handler) GetHotBytesReadStores() map[uint64]uint64 {
 }
 
 // GetHotKeysWriteStores gets all hot write stores stats.
-func (h *Handler) GetHotKeysWriteStores() map[uint64]uint64 {
+func (h *Handler) GetHotKeysWriteStores() map[uint64]float64 {
 	cluster := h.s.GetRaftCluster()
 	if cluster == nil {
 		return nil
@@ -171,7 +171,7 @@ func (h *Handler) GetHotKeysWriteStores() map[uint64]uint64 {
 }
 
 // GetHotKeysReadStores gets all hot write stores stats.
-func (h *Handler) GetHotKeysReadStores() map[uint64]uint64 {
+func (h *Handler) GetHotKeysReadStores() map[uint64]float64 {
 	cluster := h.s.GetRaftCluster()
 	if cluster == nil {
 		return nil
